+++ R/malayrootwords.R
@@ -1,7 +1,5 @@
 #' @title Data of Malay root words
-#'
-#'
-#' @format A tibble with 4365 rows and 2 variables:
+#' @format A tibble with 4295 rows and 2 variables:
 #' \describe{
 #'   \item{`Col Word`}{dbl Malay Word}
 #'   \item{`Root Word`}{dbl Malay Root Word}