+++ R/normalized.R
@@ -1,0 +1,11 @@
+#' @title Data of Malay root words
+#' @format A tibble with 153 rows and 2 variables:
+#' \describe{
+#'   \item{`Col Word`}{dbl Word}
+#'   \item{`Normalized Word`}{dbl Normalized Word}
+#' }
+
+
+
+
+"normalized"