+++ NEWS.md
@@ -89,3 +89,23 @@ sentiment_general
 10 menarik   Positive 
 # … with 1,414 more rows
 ```
+
+6. Normalized words data frame:
+
+``` r
+normalized
+# A tibble: 153 × 2
+   `Col Word` `Normalized Word`
+   <chr>      <chr>            
+ 1 ad         ada              
+ 2 ak         aku              
+ 3 akn        akan             
+ 4 ank        anak             
+ 5 ap         apa              
+ 6 awl        awal             
+ 7 bg         bagi             
+ 8 bkn        bukan            
+ 9 blm        belum            
+10 bnjr       banjir           
+# … with 143 more rows
+```